feat(ui): 完善 RichTextWidget 并添加单元测试
- 实现 RichTextWidget 文本样式渲染（粗体、删除线、下划线）
- 为 Widget 基类添加 onMouseMove 虚方法
- 添加 RichTextWidget 单元测试（41 个测试用例）
- 修复 WeatherRenderer.cpp 中 VkResult 格式化错误
- 更新事件处理器文档说明

+++ src/client/renderer/trident/weather/WeatherRenderer.cpp
@@ -391,7 +391,7 @@ void WeatherRenderer::render(VkCommandBuffer cmd,
         void* data = nullptr;
         const VkResult mapResult = vkMapMemory(m_device, m_vertexBufferMemory, 0, size, 0, &data);
         if (mapResult != VK_SUCCESS || data == nullptr) {
-            spdlog::error("WeatherRenderer: failed to map vertex buffer memory: {}", mapResult);
+            spdlog::error("WeatherRenderer: failed to map vertex buffer memory: {}", static_cast<i32>(mapResult));
             MC_TRACE_EVENT_END("rendering.weather");
             return;
         }
